--- a/engine/julia_runner_util.jl
+++ b/engine/julia_runner_util.jl
@@ -1,3 +1,0 @@
-function timed_function_call(f, input)
-    @timed f(input...)
-end