It's like ~3x faster now (and the first run isn't slow anymore). Still slower than other languages, but not sure how to make it faster, and this should be fast enough for production.

```
Submitting /home/ada/lovelace/lovelace-solutions/julia/habitable_exoplanets.jl... 4.261601 seconds.
Submitting /home/ada/lovelace/lovelace-solutions/julia/nand_gate.jl... 3.554318 seconds.
Submitting /home/ada/lovelace/lovelace-solutions/julia/rocket_science.jl... 3.668357 seconds.
Submitting /home/ada/lovelace/lovelace-solutions/julia/wind_chill.jl... 3.728901 seconds.
```

Resolves #24